refactor: 디렉터리 경로 개편
